% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_alpha.R
\name{ML_alpha}
\alias{ML_alpha}
\title{feature selection in lasso or elastic network}
\usage{
ML_alpha(mydata, method = "lasso")
}
\arguments{
\item{mydata}{the data}

\item{method}{the feature selection method,default is "lasso", "elastic" is optional}
}
\value{
test
}
\description{
feature selection in lasso or elastic network
}
\examples{
library(dplyr)
meta_dat1 <- t(meta_dat) \%>\%
  as.data.frame() \%>\%
  dplyr::mutate(group=group)
result_ML_lasso <- ML_alpha(meta_dat1,method="lasso")
}
