% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_Boruta.R
\name{ML_Boruta}
\alias{ML_Boruta}
\title{feature selection in Boruta}
\usage{
ML_Boruta(mydata)
}
\arguments{
\item{mydata}{the data}
}
\value{
test
}
\description{
feature selection in Boruta
}
\examples{
library(dplyr)
meta_dat1 <- t(meta_dat) \%>\%
  as.data.frame() \%>\%
  dplyr::mutate(group=group)
result_ML_Boruta <- ML_Boruta(meta_dat1)
}
