% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_RF.R
\name{ML_RF}
\alias{ML_RF}
\title{feature selection in random forest}
\usage{
ML_RF(mydata, ylim_min = 0, seed = NULL)
}
\arguments{
\item{mydata}{the data}

\item{ylim_min}{the min ylim,default is 0}

\item{seed}{the seed}
}
\value{
test
}
\description{
feature selection in random forest
}
\examples{
library(dplyr)
meta_dat1 <- t(meta_dat) \%>\%
  as.data.frame() \%>\%
  dplyr::mutate(group=group)
result_ML_RF <- ML_RF(meta_dat1)
result_ML_RF$p
result_ML_RF$feature_result
}
