% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_xgboost.R
\name{ML_xgboost}
\alias{ML_xgboost}
\title{feature selection in XGBoost}
\usage{
ML_xgboost(mydata)
}
\arguments{
\item{mydata}{the data}
}
\value{
test
}
\description{
feature selection in XGBoost
}
\examples{
library(dplyr)
meta_dat1 <- t(meta_dat) \%>\%
  as.data.frame() \%>\%
  dplyr::mutate(group=group)
result_ML_xgboost <- ML_xgboost(meta_dat1)
result_ML_xgboost$p
result_ML_xgboost$feature_result
}
