% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pVolcano.R
\name{pVolcano}
\alias{pVolcano}
\title{Plot the volcano figure}
\usage{
pVolcano(mydata, foldchange_threshold = 1.5)
}
\arguments{
\item{mydata}{the result of function "differential_metabolites",have column "name","fold_change",
"fdr_wilcox"}

\item{foldchange_threshold}{the fold change, default is 1.5,1/1.5}
}
\value{
test
}
\description{
Plot the volcano figure
}
\examples{
diff_result <- DM(2**meta_dat,group)
p_volcano <- pVolcano(diff_result,foldchange=1.5)
}
