% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cad_fun.R
\name{read_cad}
\alias{read_cad}
\title{Leitura do dados cadastrais}
\usage{
read_cad(file, guess_max = 30000)
}
\arguments{
\item{file}{caminho do csv2 contendo os dados cadastrais}

\item{guess_max}{estimativa do numero de observacoes da base de dados}
}
\value{
um dataframe similar ao csv2 de input com suas variaveis padroniadas
}
\description{
Esta função le os dados cadastrais. Os dados devem ser salvos em um arquivo formatdo csv e ter como delimitador o **";"** e **","** para separacao decimal. \cr
Sao 26 campos obrigatorios e devem estar ordenados conforme a sequencia abaixo: \cr
\cr
**id** - idexador do talhao \cr
**rf** - identificacao de regiao florestal \cr
**talhao** - identificacao do talhao \cr
**ciclo** - identificacao do ciclo de plantio \cr
**rotacao** - identificacao da rotacao de plantio \cr
**especie** - descricao da especie \cr
**matgen** - descricao do material genetico \cr
**espacamento** - espacamento de plantio \cr
**regime** - regime de manejo \cr
**dt_plt** - data de plantio - dd/mm/yyyy \cr
**dt_int** - data da ultima intervencao - dd/mm/yyyy \cr
**area_plt** - area plantada do talhao - ha \cr
\cr
Não importa a nomenclatura dos campos no arquivo csv, mas sim a sua posicao. \cr
O encording pre definido é o ISO-8859-1 - Latin alphabet \cr
campos adicionais podem ser adicionados apos a ultima coluna
}
\examples{

 dc <- read_cad("C:/Users/klem00125926/dados_coletor.csv")

}
