% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dc_fun.R
\name{read_dc}
\alias{read_dc}
\title{Leitura do dados de campo}
\usage{
read_dc(file, guess_max = 1e+05)
}
\arguments{
\item{file}{caminho do csv2 contendo os dados de campo do coletor}

\item{guess_max}{estimativa do numero de observacoes da base de dados}
}
\value{
um dataframe similar ao csv2 de input com suas variaveis padroniadas
}
\description{
Esta função le os dados de campo. Os dados devem ser salvos em um arquivo formatdo csv e ter como delimitador o **";"** e **","** para separacao decimal. \cr
Sao 26 campos obrigatorios e devem estar ordenados conforme a sequencia abaixo: \cr
\cr
**atividade** - descricao da atividade \cr
**rf** - identificacao de regiao florestal \cr
**talhao** - identificacao do talhao \cr
**ciclo** - identificacao do ciclo de plantio \cr
**rotacao** - identificacao da rotacao de plantio \cr
**dt_med** - data da medicao - dd/mm/yyyy \cr
**lider** - nome do lider \cr
**auxiliar** - nome do auxiliar \cr
**parcela** - dentificacao da parcela \cr
**tipo** - tipo de medicao \cr
**forma** - forma da parcela \cr
**hr_ini** - horario de inicio da medicao - HH:MM:SS \cr
**hr_fim** - horario de finalizacao da medicao - HH:MM:SS \cr
**coordX** - Coordenada geografica  \cr
**coordY** - Coordenada geografica \cr
**lado1** - Comprimento do lado da parcela -dm \cr
**lado2** - Comprimento do lado da parcela -dm \cr
**inc1** - Inclinacao da parcela - ° \cr
**inc2** - Inclinacao da parcela - ° \cr
**linha** - numero da linha \cr
**arvore** - numero da arvore na linha \cr
**cap** - circunferencia a 1,3m - mm \cr
**alt** - altura total da arvore - dm \cr
**cod1** - codigo qualitativo 1 \cr
**cod2** - codigo qualitativo 2 \cr
**codQ** - codigo qualitativo de tora \cr
\cr
Não importa a nomenclatura dos campos no arquivo csv, mas sim a sua posicao. \cr
O encording pre definido é o ISO-8859-1 - Latin alphabet \cr
}
\examples{

 dc <- read_dc("C:/Users/klem00125926/dados_coletor.csv")

}
