% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCli.R
\name{survCli}
\alias{survCli}
\title{The survival analysis and plot}
\usage{
survCli(clinical_survival)
}
\arguments{
\item{clinical_survival}{the data.frame must have the column name of status, time, group;the test data is from survival}
}
\value{
test
}
\description{
The survival analysis and plot
}
\examples{
library(survival)
names(aml)[3] ="group"
p=survCli(aml)
p
}
